Move NonStreamingProducer to common code.
